+++ app/src/desktopMain/kotlin/com/shub39/rush/core/presentation/util.desktop.kt
@@ -1,0 +1,3 @@
+package com.shub39.rush.core.presentation
+
+actual fun hypnoticAvailable() = true
